pyproject.toml
[tool.poetry]
name = "bggapi"
version = "0.1.0"
description = ""
authors = ["ellis <ellispridgeon1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,!=3.9.7,<4.0"
streamlit = "^1.26.0"
langchain = "^0.0.293"
python-dotenv = "1.0.0"
openai = "^0.28.0"
tabulate = "^0.9.0"
matplotlib = "^3.8.0"
networkx = "^3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
